Adjusting type for allocation value

+++ pkg/apis/kubermatic/v1/ipam_pool.go
@@ -61,12 +61,12 @@ type IPAMPoolDatacenterSettings struct {
 	// +kubebuilder:validation:Maximum:=128
 	// AllocationPrefix is the prefix for the allocation.
 	// Used when "type=prefix".
-	AllocationPrefix uint8 `json:"allocationPrefix,omitempty"`
+	AllocationPrefix int `json:"allocationPrefix,omitempty"`
 
 	// +kubebuilder:validation:Minimum:=1
 	// AllocationRange is the range for the allocation.
 	// Used when "type=range".
-	AllocationRange uint64 `json:"allocationRange,omitempty"`
+	AllocationRange int `json:"allocationRange,omitempty"`
 }
 
 // +kubebuilder:validation:Pattern="((^((([0-9]|[1-9][0-9]|1[0-9]{2}|2[0-4][0-9]|25[0-5])\\.){3}([0-9]|[1-9][0-9]|1[0-9]{2}|2[0-4][0-9]|25[0-5]))/([0-9]|[1-2][0-9]|3[0-2])$)|(^(([0-9a-fA-F]{1,4}:){7,7}[0-9a-fA-F]{1,4}|([0-9a-fA-F]{1,4}:){1,7}:|([0-9a-fA-F]{1,4}:){1,6}:[0-9a-fA-F]{1,4}|([0-9a-fA-F]{1,4}:){1,5}(:[0-9a-fA-F]{1,4}){1,2}|([0-9a-fA-F]{1,4}:){1,4}(:[0-9a-fA-F]{1,4}){1,3}|([0-9a-fA-F]{1,4}:){1,3}(:[0-9a-fA-F]{1,4}){1,4}|([0-9a-fA-F]{1,4}:){1,2}(:[0-9a-fA-F]{1,4}){1,5}|[0-9a-fA-F]{1,4}:((:[0-9a-fA-F]{1,4}){1,6})|:((:[0-9a-fA-F]{1,4}){1,7}|:))/([0-9]|[0-9][0-9]|1[0-1][0-9]|12[0-8])$))"
